# apps/mobile/current/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ald1n CMS Mobile API
  version: 2.2.0
  description: |
    Stabilni JSON ugovor za React Native / Expo klijente. Sve zaštićene rute
    koriste Laravel Sanctum Bearer token. Greške koriste jedinstveni envelope
    sa code i request_id poljima.
servers:
  - url: https://cms.ald1n.com
    description: Production
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/v1/auth/token:
    post:
      tags: [Auth]
      summary: Prijava i izdavanje Sanctum tokena
      security: []
      operationId: createAuthToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login, password, device_name]
              properties:
                login: { type: string, maxLength: 190 }
                password: { type: string, maxLength: 4096 }
                device_name: { type: string, maxLength: 120 }
      responses:
        '201':
          description: Token je izdat
          content:
            application/json:
              schema:
                type: object
                required: [token, token_type, user, permissions, api_version]
                properties:
                  token: { type: string }
                  token_type: { type: string, const: Bearer }
                  expires_at: { type: [string, 'null'], format: date-time }
                  user: { $ref: '#/components/schemas/UserSummary' }
                  permissions:
                    type: array
                    items: { type: string }
                  api_version: { type: string, const: v1 }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
    delete:
      tags: [Auth]
      summary: Opoziv trenutnog tokena i vezanog uređaja
      operationId: revokeCurrentAuthToken
      responses:
        '204': { description: Token je opozvan }
        '401': { $ref: '#/components/responses/Unauthenticated' }
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
  schemas:
    UserSummary:
      type: object
      required: [id, username, name, status]
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: [string, 'null'], format: email }
        name: { type: string }
        role: { type: [string, 'null'] }
        group: { type: [string, 'null'] }
        status: { type: string }
...
    ApiError:
      type: object
      required: [message, code, request_id]
      properties:
        message: { type: string }
        code: { type: string }
        request_id: { type: [string, 'null'] }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
  responses:
    ValidationError:
      description: Validacija nije prošla
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthenticated:
      description: Token nedostaje, istekao je ili je opozvan
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
    RateLimited:
      description: Prekoračeno ograničenje zahteva
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }